rearranging settings variables
Introduced a variable for the Azure Top Level Domain ($azureTLD) to the switch statement that currently picks the correct Scope/Token URLs based on the chosen Azure Cloud. This switch was then moved up to immediately follow the connectivity settings variables near the beginning of the script. In doing so, the logic was removed for determining if Exchange Online is used.

As a result, these variables are now always set/made available so it's possible to drive connectivity to either Exchange Online, Azure Cognitive Services, both, or none.