Support for Azure Government (#356)
* feature icon

azure government icon

* feature, azure government

brief description for Azure Government connectivity

* link text

updating per feature

* Azure Cloud Instance enum

Adding an enum for Azure Cloud instances to choose from when defining connection to Exchange

* update ordinal

changing ordinal for sorting in the ui

* load method

loads the Azure Cloud Instance enums and defined enum (if any) for the chosen azure cloud instance

* save method

save the chosen Azure Cloud Instance to the settings mp

* changing property name

removing "enum" from the property name to load/set on the mp

* renaming property

following naming practice within the core admin settings class

* Azure Cloud Instance picker

combobox to select which Azure Cloud Instance should be used to define the connection to Exchange Online

* switch, azure cloud instance

if exchange online is being used, pick the correct Scope and Token URLs based on the chosen Azure Cloud Instance. default to Azure Public.

* variable rename

renaming variable to provide a bit more standardization

* rearranging settings variables

Introduced a variable for the Azure Top Level Domain ($azureTLD) to the switch statement that currently picks the correct Scope/Token URLs based on the chosen Azure Cloud. This switch was then moved up to immediately follow the connectivity settings variables near the beginning of the script. In doing so, the logic was removed for determining if Exchange Online is used.

As a result, these variables are now always set/made available so it's possible to drive connectivity to either Exchange Online, Azure Cognitive Services, both, or none.

* moving cloud instance picker

Since the Azure Cloud Instance picker is used to drive Exchange and/or Azure AI services it's visibility is always required as opposed to being dependent on whether or not Exchange Online is used

* revert last commit

rather than make this a seemingly mandatory config on the general page, it should be kept part of 365 connectivity

* translation, azure cloud instance

in the event only Azure Translation Services are being used, provide a cloud instance picker

* vision, azure cloud instance

in the event only Azure Vision Services are being used, provide a cloud instance picker

* transcription, azure cloud instance

in the event only Azure Speech Services are being used, provide a cloud instance picker

* sentiment, azure cloud instance

in the event only Azure Sentiment Analysis is being used or used in conjunction with machine learning, provide a cloud instance picker