v5.0.3 (#457)
* Fix for Templates with nested Activities (#447)

Added -Alias $alias to the recursive Update-SCSMPropertyCollection call

* Fix for Update-WorkItem with Comment length gt 4000

Whenever the update-workitem is called with a message, where the length of the body is greater than 4000, the substring(0,4000) function is called, but the result is not written back to the commentToAdd variable - so the full message is further processed, and the action log is not added to the workitem.

* Replace the ([R][E][:]) regex for Message merging with a multi language Regex (#453)

* Update smletsExchangeConnector.ps1

* Update smletsExchangeConnector.ps1

* update reply regex

adding square brackets to two of the reply regexes to be consistent with the rest of the patterns

---------

Co-authored-by: Adam

* move true/false value when loading Custom Events (#455)

$true/$false values are incorrectly tied to the logging level used by the connector. This moves them out of that logic and ensures $ceScripts always has a value

* Warning, PSScriptAnalyzer

Per PSScriptAnalyzer, this specific regex reply pattern is throwing a warning. It's possible this can be overcome, but short of further research/holding the release it's going to be removed for now. https://learn.microsoft.com/en-us/powershell/utility-modules/psscriptanalyzer/rules/usebomforunicodeencodedfile?view=ps-modules

* contributors (#458)

* powershell contributors

updating powershell contributor list per pull requests/commits

* update contributors on AboutForm.xaml

updating the contributor list per pull requests/commits

* version notes (#459)

updating inline version notes with relevant issues/pull requests

---------

Co-authored-by: SimonZein <128088337+SimonZein@users.noreply.github.com>