MA: notes + comment to add (#498)
* notes + comment to add

initial commit, meets parity with stock connector to prevent the comment + current notes from exceeding 4,000 characters. introduces a Custom Event to override this behavior.

* manaul activity note event

introduces an event/function when the manual activity notes are greater than 4000 characters

* logic refactor

rather than update $commentToAdd with the current notes to calculate/use a single variable. This performs the math within the IF statement. In this way, $commentToAdd and the original Notes are kept separate so if there is need to override functionality with the Custom Event that can be done

* redundancy

Work Item Notes existed twice in the character count check

* case

should be uppercase

* event verbiage

updating the event message to be more accurate of what is happening